backend/pyproject.toml
[project]
name = "linguistic-backend"
version = "0.1.0"
description = "Linguistic Analysis Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.28",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.5.0",
    "uvicorn>=0.35.0",
]
